Change reported DPI to correct value [320 -> 359]
DPI in original file (from stock OS) was 320 so DPI-specific GApps would operate. 
Actual DPI is 359. 
Do the math, if you like. Or just look it up. But objects displayed on the screen in CM are too small when scaled to 320DPI. Just a cosmetics thing. 
Of course, DPI can be changed in [Settings -> Display & lights -> LCD density] or by editing the build.prop file and rebooting. 
By the way, I will propose the same change to build.prop.